requirements.txt
pdfplumber 
sentence-transformers 
faiss-cpu 
numpy 
pandas 
beautifulsoup4